requirements.txt
dj-database-url==0.4.2
Django==1.11.7
django-appconf==1.0.2
django-autocomplete-light==3.2.9
django-autoslug==1.9.3
django-bootstrap3==8.2.2
django-braces==1.11.0
django-cities-light==3.3.0
django-debug-toolbar==1.7
django-extra-views==0.9.0
django-imagekit==4.0.1
django-localflavor==1.5.2
django-markdown2==0.3.0
django-model-utils==3.0.0
django-smart-selects==1.5.2
django-static-jquery==2.1.4
django-tables2==1.10.0
djangorestframework==3.6.3
gunicorn==19.7.1
markdown2==2.3.3
pilkit==2.0
Pillow==2.6.1
progressbar2==3.30.2
psycopg2==2.7.3.1
pypandoc==1.4
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.21
